@@ -1,10 +1,12 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.6.1'
+ruby '2.7.2'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
-gem 'rails', '~> 6.1.3', '>= 6.1.3.1'
+gem "rails", "~> 7.0.0"
+#uncomment to fix gem dependencies issues from upgrading ruby and rails versions
+gem "net-http"
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 1.1'
 gem 'rack-mini-profiler'
